requirements.txt
boto3
streamlit
extra-streamlit-components
pyyaml
